Houseclassificationapp.py
import streamlit as st
import tensorflow as tf
from PIL import Image
import numpy as np

# Load the model and labels
model = tf.keras.models.load_model('keras_model.h5')
with open('labels.txt', 'r') as file:
    class_labels = file.read().splitlines()

# Define image classification function
def classify_image(image_path):
    img = Image.open(image_path).resize((224, 224))
    img_array = np.array(img) / 255.0
    img_array = np.expand_dims(img_array, axis=0)
    prediction = model.predict(img_array)[0]
    class_index = np.argmax(prediction)
    confidence = prediction[class_index]
    class_label = class_labels[class_index]
    return class_label, confidence

# Create Streamlit app
st.title("House Classification App")

uploaded_file = st.file_uploader("Choose a file...", type=["jpg", "jpeg", "png"])

if uploaded_file is not None:
    st.image(uploaded_file, caption="Uploaded Image.", use_column_width=True)
    st.write("")
    st.write("Classifying...")

    # Perform inference
    class_label, confidence = classify_image(uploaded_file)
    st.write(f"Predicted Class: {class_label}")
    st.write(f"Confidence Score: {confidence:.2%}")
