requirements.txt
numpy>=1.15.1
Pillow>=9.0.1
streamlit==1.14.0  
tensorflow==2.5.0

